Move PeerInterconnect into its own module
Also moves the PeerIntPeerInterconnect
