+++ libsplinter/src/peer/interconnect/error.rs
@@ -1,0 +1,34 @@
+// Copyright 2018-2022 Cargill Incorporated
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+use std::{error, fmt};
+
+/// Errors that could be raised by `PeerInterconnect`
+#[derive(Debug, PartialEq)]
+pub enum PeerInterconnectError {
+    /// `PeerInterconnect` start up failed
+    StartUpError(String),
+}
+
+impl error::Error for PeerInterconnectError {}
+
+impl fmt::Display for PeerInterconnectError {
+    fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
+        match self {
+            PeerInterconnectError::StartUpError(msg) => {
+                write!(f, "Unable to start peer interconnect: {}", msg)
+            }
+        }
+    }
+}
